Added fundable.rb Moved funding logic into a module
